5(Audio)/requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Data processing
PyPDF2>=3.0.0
python-docx>=0.8.11
nltk>=3.8.1
spacy>=3.6.0
beautifulsoup4>=4.12.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
pyttsx3>=2.90
gtts>=2.3.0
pygame>=2.5.0
ffmpeg-python>=0.2.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
numpy>=1.22.4,<2.3.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pywin32>=306

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
